# openapi.yaml
openapi: 3.1.0
info:
  title: arca API
  version: v1
servers:
- url: https://arca-backend.dcism.org
  description: Production
- url: http://localhost:20255
  description: Local development
paths:
  /api/v1/vaults/edit-label:
    put:
      tags:
      - vault-controller
      operationId: editLabel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditVaultLabelRequest"
        required: true
      responses:
        "200":
          description: Label updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VaultResponse"
  /api/v1/user/{userId}:
    get:
      tags:
      - user-controller
      operationId: getUserDetails
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
      - user-controller
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
        required: true
      responses:
        "200":
          description: User updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      tags:
      - user-controller
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/user/{userId}/profile-picture:
    get:
      tags:
      - user-controller
      operationId: getProfilePicture
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Profile picture
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Profile picture not found
    put:
      tags:
      - user-controller
      operationId: updateProfilePicture
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: Profile picture updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfilePictureResponse"
...
  /api/v1/vaults/add:
    post:
      tags:
      - vault-controller
      operationId: addToVault
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultRequest"
        required: true
      responses:
        "200":
          description: Post saved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/user:
    get:
      tags:
      - user-controller
      operationId: getAllUser
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
    post:
      tags:
      - user-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
  /api/v1/user/me:
    get:
      tags:
      - user-controller
      operationId: getCurrentUser
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    patch:
      tags:
      - user-controller
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /api/v1/vaults/{id}:
    get:
      tags:
      - vault-controller
      operationId: getVaultEntry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Vault entry
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VaultResponse"
...
  /api/v1/vaults/check/{postId}:
    get:
      tags:
      - vault-controller
      operationId: isPostSaved
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Saved state
          content:
            '*/*':
              schema:
                type: boolean
  /api/v1/user/search:
    get:
      tags:
      - user-controller
      operationId: searchUsers
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummaryResponse"
  /api/v1/posts/user/{userId}:
    get:
      tags:
      - post-controller
      operationId: getAllUserPost
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"
...
  /api/v1/admin/users/{id}:
    get:
      tags:
      - admin-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
      - admin-controller
      operationId: deleteUser_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /api/v1/vaults/remove:
    delete:
      tags:
      - vault-controller
      operationId: removeFromVault
      parameters:
      - name: postId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Post removed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
    EditVaultLabelRequest:
      type: object
      properties:
        postId:
          type: integer
          format: int64
        newLabel:
          type: string
    VaultResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        post_id:
          type: integer
          format: int64
    UserUpdateRequest:
      type: object
      properties:
        bio:
          type: string
        course:
          type: string
        department:
          type: string
        first_name:
          type: string
        last_name:
          type: string
    ProfilePictureResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        profilePicture:
          type: string
        message:
          type: string
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    VaultRequest:
      type: object
      properties:
        postId:
          type: integer
          format: int64
        label:
          type: string
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        course:
          type: string
        department:
          type: string
        bio:
          type: string
        roles:
          type: array
          items:
            type: string
        first_name:
          type: string
        last_name:
          type: string
        profile_picture:
          type: string
    UserSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
        course:
          type: string
        department:
          type: string
        bio:
          type: string
        profile_picture:
          type: string
    FileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
          format: int64
        downloadUrl:
          type: string
    PostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        postId:
          type: integer
          format: int32
        version:
          type: integer
          format: int32
        title:
          type: string
        content:
          type: string
        status:
          type: string
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        departmentId:
          type: integer
          format: int64
        departmentName:
          type: string
        postTag:
          type: string
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/PostSubjectResponse"
        rejectionReason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        upvoteCount:
          type: integer
          format: int32
        downvoteCount:
          type: integer
          format: int32
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileResponse"
        is_latest_version:
          type: boolean
    PostSubjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...